refactor: notification2 패키지명을 notification로 변경
