+++ src/main/java/com/sesac/joinflex/domain/friend/service/FriendRequestService.java
@@ -5,8 +5,8 @@ import com.sesac.joinflex.domain.friend.dto.response.FriendResponse;
 import com.sesac.joinflex.domain.friend.entity.FriendRequest;
 import com.sesac.joinflex.domain.friend.entity.FriendRequestStatus;
 import com.sesac.joinflex.domain.friend.repository.FriendRequestRepository;
-import com.sesac.joinflex.domain.notification2.message.InviteMessageTemplate;
-import com.sesac.joinflex.domain.notification2.service.NotificationService;
+import com.sesac.joinflex.domain.notification.message.InviteMessageTemplate;
+import com.sesac.joinflex.domain.notification.service.NotificationService;
 import com.sesac.joinflex.domain.user.entity.User;
 import com.sesac.joinflex.domain.user.repository.UserRepository;
 import com.sesac.joinflex.global.exception.CustomException;
